LPS-98626 Move *TestCase classes of the package "om.liferay.portlet.subscriptions.test" to subscription-test-util
test
