+++ modules/apps/subscription/subscription-test-util/src/main/java/com/liferay/subscription/test/util/BaseSubscriptionLocalizedContentTestCase.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.subscriptions.test;
+package com.liferay.subscription.test.util;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.model.Layout;
