+++ portal-test-integration/bnd.bnd
@@ -22,7 +22,6 @@ Export-Package:\
 	com.liferay.portlet.documentlibrary.util.test,\
 	com.liferay.portlet.notifications.test,\
 	com.liferay.portlet.social.test,\
-	com.liferay.portlet.subscriptions.test,\
 	com.liferay.portlet.trash.test
 Import-Package: *;resolution:=optional
 Include-Resource:\
