--- portal-test-integration/src/com/liferay/portlet/subscriptions/test/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
